+++ flink-mesos/src/main/java/org/apache/flink/mesos/util/MesosUtils.java
@@ -36,7 +36,7 @@ import org.apache.flink.runtime.clusterframework.overlays.KeytabOverlay;
 import org.apache.flink.runtime.clusterframework.overlays.Krb5ConfOverlay;
 import org.apache.flink.runtime.clusterframework.overlays.SSLStoreOverlay;
 import org.apache.flink.runtime.clusterframework.overlays.UserLibOverlay;
-import org.apache.flink.runtime.util.ClusterEntrypointUtils;
+import org.apache.flink.runtime.entrypoint.ClusterEntrypointUtils;
 
 import org.apache.mesos.Protos;
 import org.slf4j.Logger;
